baseline_experiment/requirements.txt
# RAPTOR baseline 의존성
# 기본 의존성은 프로젝트 루트 requirements 에 이미 포함되어 있습니다.
# RAPTOR 전용 추가 패키지:

umap-learn>=0.5.6
scikit-learn>=1.3.0
tiktoken>=0.7.0
PyMuPDF>=1.24.0
openai>=1.30.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
